Rename legacy/legacyRunner to legacy/legacy-query-runner (kebab-case and class-file name)
