+++ extensions/ql-vscode/src/query-server/legacy/legacy-query-runner.ts
